+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/StringUUID.kt
@@ -1,0 +1,5 @@
+package com.narbase.narcore.dto.common
+
+
+typealias StringUUID = String
+fun Any.toStringUUID(): StringUUID = this.toString()
