+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/network/BasicResponse.kt
@@ -1,4 +1,6 @@
-package com.narbase.narcore.web.network
+package com.narbase.narcore.dto.common.network
+
+import kotlin.js.JsExport
 
 /**
  * Created by Me on 17/02/2017.
