+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/network/ItemList.kt
@@ -1,4 +1,6 @@
-package com.narbase.narcore.web.network.dto
+package com.narbase.narcore.dto.common.network
+
+import kotlin.js.JsExport
 
 /*
  * Copyright 2017-2020 Narbase technologies and contributors. Use of this source code is governed by the MIT License.
