+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/utils/MultiLingualTextDto.kt
@@ -1,0 +1,6 @@
+package com.narbase.narcore.dto.common.utils
+
+import kotlin.js.JsExport
+
+@JsExport
+data class MultiLingualTextDto(val en: String, val ar: String)
