+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/utils/files/CreateFileDtos.kt
@@ -1,0 +1,8 @@
+package com.narbase.narcore.dto.common.utils.files
+
+object CreateFileDtos {
+    class ResponseDto(
+        val url: String,
+        val fileName: String
+    )
+}
