+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/common/utils/uploaders/UploadFileResponseDto.kt
@@ -1,0 +1,9 @@
+package com.narbase.narcore.dto.common.utils.uploaders
+
+import kotlin.js.JsExport
+
+@JsExport
+class UploadFileResponseDto(
+    val url: String,
+    val fileName: String
+)
