+++ dto-web/src/commonMain/kotlin/com/narbase/narcore/dto/domain/user/profile/UpdateUserProfileDto.kt
@@ -1,4 +1,6 @@
-package com.narbase.narcore.web.views.user.profile
+package com.narbase.narcore.dto.domain.user.profile
+
+import kotlin.js.JsExport
 
 /*
  * Copyright 2017-2020 Narbase technologies and contributors. Use of this source code is governed by the MIT License.
