+++ narcore-server/src/main/kotlin/com/narbase/narcore/common/UUID.kt
@@ -1,0 +1,5 @@
+package com.narbase.narcore.common
+
+import java.util.*
+
+fun String.toUUID(): UUID = UUID.fromString(this)
