+++ narcore-server/src/main/kotlin/com/narbase/narcore/core/MultiLingualText.kt
@@ -1,12 +1,12 @@
 package com.narbase.narcore.core
 
+import com.narbase.narcore.dto.common.utils.MultiLingualTextDto
+
 
 data class MultiLingualText(val en: String, val ar: String) {
     fun toDto(): MultiLingualTextDto = MultiLingualTextDto(en, ar)
 }
 
-data class MultiLingualTextDto(val en: String, val ar: String) {
-    fun toDs(): MultiLingualText = MultiLingualText(en, ar)
-}
+fun MultiLingualTextDto.toDs(): MultiLingualText = MultiLingualText(en, ar)
