+++ narcore-server/src/main/kotlin/com/narbase/narcore/domain/user/profile/UpdateProfileController.kt
@@ -5,15 +5,16 @@ import com.narbase.narcore.common.Handler
 import com.narbase.narcore.common.auth.loggedin.AuthorizedClientData
 import com.narbase.narcore.common.exceptions.UnauthenticatedException
 import com.narbase.narcore.data.tables.UsersTable
+import com.narbase.narcore.dto.domain.user.profile.UpdateUserProfileDto
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.jetbrains.exposed.sql.update
 import java.util.*
 
-class UpdateProfileController : Handler<UpdateProfileController.RequestDto, Unit>(
-    RequestDto::class
+class UpdateProfileController : Handler<UpdateUserProfileDto.RequestDto, Unit>(
+    UpdateUserProfileDto.RequestDto::class
 ) {
     @Suppress("RemoveRedundantQualifierName")
-    override fun process(requestDto: RequestDto, clientData: AuthorizedClientData?): DataResponse<Unit> {
+    override fun process(requestDto: UpdateUserProfileDto.RequestDto, clientData: AuthorizedClientData?): DataResponse<Unit> {
         val clientId = UUID.fromString(clientData?.id ?: throw UnauthenticatedException())
         transaction {
             UsersTable.update({ UsersTable.clientId eq clientId }) {
@@ -24,12 +25,4 @@ class UpdateProfileController : Handler<UpdateProfileController.RequestDto, Unit
         }
         return DataResponse()
     }
-
-    class RequestDto(
-        val fullName: String,
-        val callingCode: String,
-        val localPhone: String
-    )
-
-    class ResponseDto()
 }
