+++ narcore-web/src/jsMain/kotlin/com/narbase/narcore/web/network/calls/settings/AdminStaffServerCaller.kt
@@ -1,9 +1,9 @@
 package com.narbase.narcore.web.network.calls.settings
 
-import com.narbase.narcore.data.dto.roles.DynamicRoleDto
+import com.narbase.narcore.dto.domain.admin.AdminStaffServerCallerDtos
 import com.narbase.narcore.web.network.ServerCaller
 import com.narbase.narcore.web.network.crud.CrudServerCaller
-import com.narbase.narcore.web.utils.DataResponse
+import com.narbase.narcore.dto.common.network.DataResponse
 
 /*
  * Copyright 2017-2020 Narbase technologies and contributors. Use of this source code is governed by the MIT License.
@@ -21,24 +21,3 @@ object AdminStaffServerCaller :
 
 }
 
-@JsExport
-object AdminStaffServerCallerDtos {
-    @Suppress("unused")
-    class Filters(
-        val getInactive: Boolean?,
-        val clientId: String?
-    )
-
-
-    @Suppress("unused")
-    class StaffDto(
-        val clientId: String?,
-        val userId: String?,
-        val username: String,
-        val password: String,
-        val fullName: String,
-        val callingCode: String,
-        val localPhone: String,
-        val dynamicRoles: Array<DynamicRoleDto>
-    )
-}
