+++ narcore-web/src/jsMain/kotlin/com/narbase/narcore/web/network/dto/roles/DynamicRoleDto.kt
@@ -1,5 +1,6 @@
 package com.narbase.narcore.data.dto.roles
 
+import com.narbase.narcore.dto.models.roles.DynamicRoleDto
 import com.narbase.narcore.dto.models.roles.Privilege
 import com.narbase.narcore.web.network.valueOfDto
 
@@ -8,15 +9,6 @@ import com.narbase.narcore.web.network.valueOfDto
  */
 
 typealias PrivilegeName = String
-@JsExport
-data class DynamicRoleDto(
-    val id: String?,
-    val name: String,
-    val privileges: Array<PrivilegeName>,
-    val isDoctor: Boolean
-) {
-
-}
 
 val DynamicRoleDto.privilegesEnums
     get() = privileges.mapNotNull { valueOfDto<Privilege>(it) }
