+++ narcore-web/src/jsMain/kotlin/com/narbase/narcore/web/views/admin/staff/StaffManagementViewModel.kt
@@ -4,11 +4,11 @@ import com.narbase.kunafa.core.lifecycle.Observable
 import com.narbase.narcore.web.network.UnknownErrorException
 import com.narbase.narcore.web.network.basicNetworkCall
 import com.narbase.narcore.web.network.calls.settings.AdminStaffServerCaller
-import com.narbase.narcore.web.network.calls.settings.AdminStaffServerCallerDtos
+import com.narbase.narcore.dto.domain.admin.AdminStaffServerCallerDtos
 import com.narbase.narcore.web.network.calls.settings.EnableUserDto
-import com.narbase.narcore.web.network.crud.CrudDto
+import com.narbase.narcore.dto.common.network.crud.CrudDto
 import com.narbase.narcore.web.utils.BasicUiState
-import com.narbase.narcore.web.utils.DataResponse
+import com.narbase.narcore.dto.common.network.DataResponse
 
 /*
  * Copyright 2017-2020 Narbase technologies and contributors. Use of this source code is governed by the MIT License.
